pyproject.toml
[tool.poetry]
name = "hm310p-cli"
version = "0.1.0"
description = "Command line tool to control Hanmatek power supplies."
license = "MIT"
readme = "README.md"
homepage = "https://github.com/5H3LL3H5/hm310p-cli"
repository = "https://github.com/5H3LL3H5/hm310p-cli"
keywords = ["hm305p", "hm310p", "hm305p-cli", "hm310p-cli"]
authors = ["Christian Stenzel <christianstenzel@linux.com>"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.1.2"
minimalmodbus = "^1.0.2"
pyserial = "^3.4"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
coverage = {extras = ["toml"], version = "^5.3"}
pytest-cov = "^2.10.1"
pytest-mock = "^3.3.1"
black = "^20.8b1"
flake8 = "^3.8.4"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.1"
flake8-bugbear = "^20.1.4"
flake8-import-order = "^0.18.1"
safety = "^1.9.0"
mypy = "^0.790"
pytype = {version = "^2020.1.8", python = "3.7"}
flake8-annotations = "^2.0.0"
typeguard = "^2.7.1"
flake8-docstrings = "^1.5.0"
darglint = "^1.1.2"
xdoctest = "^0.11.0"
sphinx = "^2.3.1"
sphinx-autodoc-typehints = "^1.10.3"
codecov = "^2.0.22"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
hm310p-cli = "hm310p_cli.console:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["hm310p_cli"]

[tool.coverage.report]
show_missing = true
fail_under = 100
